Add files via upload
The purpose of this project is to analyze the performance of existing solar dryer present in BEC renewable energy park. The solar drying system utilizes solar energy to heat up air and to dry any food substance loaded, which is beneficial in increasing shelf life of food product thereby reducing the wastage of agricultural product and helps in preservation of agricultural product. 

Drying is an eminent way to preserve the food and solar energy food drying is an approximate food preservation mechanism for a sustainable real world. This fixed solar dryer has the capacity of 5 kg which is used for the preservation, drying of grapes, potatoes, chilies, peas, ground nut, coconut, sweet corn, sun flower seeds, tur dal etc., more than 10 kinds have been dried using this solar dryer. Drying will generally refer to the removal of moisture content by evaporation rather than by pressure or other physical parameters. Our country is blessed with ample of solar energy round the year. The principle of this dryer is that, hot air is lighter than the cool air and its raises up the altitude. While raising this warm air comes in contact with food slices and draws the moisture from it. The repeated cycle of this process makes it a low cost, very healthy, long term investment. Generally, the sun’s power of heat is used to dry up the moisture content of the fruits or vegetables.

The challenge is to modify the dryer utilizing the abundant solar radiation, provided the efficiency of the dryer is not compromised. Also it should have better temperature regulation and extended operational hours. With this intent, we are proposing a study of the performance of Tray Dryer.